requirements.txt
allure-python-commons
allure-pytest==2.15.2
pytest==9.0.2
requests==2.32.5
jsonschema==4.25.1
python-dotenv==1.2.1